agro_backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.2
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0
numpy==1.26.4
pillow==10.0.0
python-multipart==0.0.6
aiofiles==23.1.0
boto3==1.29.0
requests==2.31.0
python-dotenv==1.0.0